fix(go): model fidelity — enabled default true, score 1.0/0.0 JSON
Two parity fixes found in the pre-cutover sweep (docs/SWEEP.md):

- CheckDefinition.enabled now defaults to true when the YAML key is
  absent (custom UnmarshalYAML), matching pydantic's `enabled: bool =
  True`. Previously Go's zero value silently disabled checks in any
  custom policy that omitted the key.
- New models.Score type marshals integral scores as 1.0/0.0 (matching
  pydantic) instead of encoding/json's 1/0. Fires on every perfect or
  zero score; the numeric differential test could not catch it.

+++ internal/actions/github_issues.go
@@ -134,7 +134,7 @@ var statusIcons = map[models.CheckStatus]string{
 // BuildBody renders the markdown issue body. Exported for golden testing.
 func BuildBody(result models.RepoResult, now time.Time) string {
 	timestamp := now.UTC().Format("2006-01-02 15:04 UTC")
-	scorePct := fmt.Sprintf("%.0f%%", result.Score*100)
+	scorePct := fmt.Sprintf("%.0f%%", float64(result.Score)*100)
 
 	rows := make([]string, 0, len(result.Results)+2)
 	rows = append(rows, "| check | status | severity | message |", "|---|---|---|---|")
